backend/requirements.txt
django==4.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
faiss-cpu==1.8.0.post1
fasttext-wheel==0.9.2
scikit-learn==1.1.3
torch==2.3.1
tqdm==4.66.4
transformers==4.37.0
openpyxl==3.1.5
pandas==2.2.2
protobuf==3.19.5
pillow==10.4.0
nltk==3.8.1
numpy==1.23.5
keras==2.8.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
SQLAlchemy==1.4.52
tensorflow==2.8.4
python-dotenv==1.0.1